+++ XliffSync/Model/XlfDocument.ps1
@@ -155,8 +155,7 @@ class XlfDocument {
                 return $this.idUnitMap[$transUnitId];
             }
             return $null;
-        }
-        else {
+        } else {
             return $this.TranslationUnitNodes() | Where-Object { $_.'id' -eq $transUnitId } | Select-Object -First 1;
         }
     }
@@ -168,8 +167,7 @@ class XlfDocument {
                 return $this.xliffGeneratorNoteSourceUnitMap[$key];
             }
             return $null;
-        }
-        else {
+        } else {
             return $this.TranslationUnitNodes() | Where-Object { ($this.GetUnitXliffGeneratorNote($_) -eq $xliffGenNote) -and ($this.GetUnitSourceText($_) -eq $sourceText) } | Select-Object -First 1;
         }
     }
@@ -181,8 +179,7 @@ class XlfDocument {
                 return $this.xliffGeneratorNoteDeveloperNoteUnitMap[$key];
             }
             return $null;
-        }
-        else {
+        } else {
             return $this.TranslationUnitNodes() | Where-Object { ($this.GetUnitXliffGeneratorNote($_) -eq $xliffGenNote) -and ($this.GetUnitDeveloperNote($_) -eq $devNote) } | Select-Object -First 1;
         }
     }
@@ -194,8 +191,7 @@ class XlfDocument {
                 return $this.xliffGeneratorNoteUnitMap[$key];
             }
             return $null;
-        }
-        else {
+        } else {
             return $this.TranslationUnitNodes() | Where-Object { ($this.GetUnitXliffGeneratorNote($_) -eq $xliffGenNote) } | Select-Object -First 1;
         }
     }
@@ -207,8 +203,7 @@ class XlfDocument {
                 return $this.sourceDeveloperNoteUnitMap[$key];
             }
             return $null;
-        }
-        else {
+        } else {
             return $this.TranslationUnitNodes() | Where-Object { ($this.GetUnitDeveloperNote($_) -eq $devNote) -and ($this.GetUnitSourceText($_) -eq $sourceText) } | Select-Object -First 1;
         }
     }
@@ -220,8 +215,7 @@ class XlfDocument {
                 return $this.sourceUnitMap[$key];
             }
             return $null;
-        }
-        else {
+        } else {
             return $this.TranslationUnitNodes() | Where-Object { ($this.GetUnitSourceText($_) -eq $sourceText) } | Select-Object -First 1;
         }
     }
@@ -285,8 +279,7 @@ class XlfDocument {
                             $sourceUnitAsElement.SetAttribute($attr.Name, $attr.Value);
                         }
                     }
-                }
-                else {
+                } else {
                     foreach ($attr in $targetUnit.Attributes) {
                         if ($attr.Name -ne 'id') {
                             $newAttr = $this.root.OwnerDocument.ImportNode($attr, $true);
@@ -294,8 +287,7 @@ class XlfDocument {
                         }
                     }
                 }
-            }
-            else {
+            } else {
                 # Use the source's attribute values for the attributes in common, and extend these with any extra attributes from the target.
                 foreach ($attr in $targetUnit.Attributes) {
                     if (-not $sourceUnitAsElement.Attributes[$attr.Name]) {
@@ -319,8 +311,7 @@ class XlfDocument {
                 $newTranslationState = [XlfTranslationState]::MissingTranslation;
             }
             $targetNode = $this.CreateTargetNode($sourceUnit, $translation, $newTranslationState);
-        }
-        elseif ((-not $needsTranslation) -and $targetNode) {
+        } elseif ((-not $needsTranslation) -and $targetNode) {
             $this.DeleteTargetNode($sourceUnit);
         }
 
@@ -389,8 +380,7 @@ class XlfDocument {
         if ((-not $stateNode) -and ($this.Version() -eq "1.2")) {
             [System.Xml.XmlNode] $newTargetNode = $this.CreateTargetNode($unit, "", $newTranslationState);
             $this.AppendTargetNode($unit, $newTargetNode);
-        }
-        elseif ($stateNode) {
+        } elseif ($stateNode) {
             $this.UpdateStateAttributes($stateNode, $newTranslationState);
         }
     }
@@ -432,15 +422,13 @@ class XlfDocument {
 
         if ($existingNote) {
             $notesParent.ReplaceChild($noteNode, $existingNote);
-        }
-        elseif ($targetChildNode -and ($this.Version() -eq "1.2")) {
+        } elseif ($targetChildNode -and ($this.Version() -eq "1.2")) {
             $unit.InsertAfter($noteNode, $targetChildNode.NextSibling);
 
             # Add the same whitespace after the note.
             $newWhiteSpaceNode = $this.root.OwnerDocument.ImportNode($targetChildNode.PreviousSibling, $true);
             $unit.InsertAfter($newWhiteSpaceNode, $noteNode);
-        }
-        else {
+        } else {
             $notesParent.AppendChild($noteNode);
         }
     }
@@ -507,20 +495,17 @@ class XlfDocument {
 
                 if ($targetChildNode) {
                     $unit.ReplaceChild($targetNode, $targetChildNode);
-                }
-                elseif ($sourceChildNode) {
+                } elseif ($sourceChildNode) {
                     if ($unit.Attributes["xml:space"] -and ($unit.Attributes["xml:space"].Value -eq "preserve")) {
                         $unit.InsertAfter($targetNode, $sourceChildNode.NextSibling);
 
                         # Add the same whitespace after the target node.
                         $newWhiteSpaceNode = $this.root.OwnerDocument.ImportNode($sourceChildNode.PreviousSibling, $true);
                         $unit.InsertAfter($newWhiteSpaceNode, $targetNode);
-                    }
-                    else {
+                    } else {
                         $unit.InsertAfter($targetNode, $sourceChildNode);
                     }
-                }
-                else {
+                } else {
                     $unit.AppendChild($targetNode);
                 }
                 break;
@@ -573,7 +558,7 @@ class XlfDocument {
         [System.Xml.XmlNode] $noteNode = $null;
 
         switch ($this.Version()) {
-            "1.2" { 
+            "1.2" {
                 $noteNode = $unitNode.ChildNodes | Where-Object { ($_.Name -eq "note") -and ($_.'from' -eq $noteDesignation) } | Select-Object -First 1;
                 break;
             }
@@ -623,7 +608,7 @@ class XlfDocument {
         }
 
         switch ($this.Version()) {
-            "1.2" { 
+            "1.2" {
                 [System.Xml.XmlNode] $bodyNode = [XlfDocument]::GetNode('body', $this.root);
                 if ($bodyNode) {
                     $unitsInBody = $this.GetTranslationUnitsFromRoot($bodyNode);
@@ -646,7 +631,9 @@ class XlfDocument {
     }
 
     [void] SaveToFilePath([string] $filePath) {
-        $filePath = Resolve-Path $filePath
+        if (Test-Path $filePath) {
+            $filePath = Resolve-Path $filePath
+        }
         $this.root.OwnerDocument.Save($filePath);
     }
 
@@ -687,8 +674,7 @@ class XlfDocument {
 
         if ($node.Name -eq $tag) {
             return $node;
-        }
-        else {
+        } else {
             foreach ($member in $node.ChildNodes) {
                 [System.Xml.XmlNode] $child = $member;
                 $reqNode = [XlfDocument]::GetNode($tag, $child);
@@ -744,7 +730,7 @@ class XlfDocument {
         $newXmlDoc.AppendChild($newRootNode);
 
         switch ($baseXlfDoc.Version()) {
-            "1.2" { 
+            "1.2" {
                 $baseFileNode = $rootNode.ChildNodes.Item(0);
                 $newFileNode = $newXmlDoc.ImportNode($baseFileNode, $false);
                 $newFileNode.'target-language' = $language;